Add QPepperInstance::runJavascript()
This allows Qt to inject and run javascript on the
hosting web page. Add a minimal "bootstrap" message
handler to the js generated by nacldeployqt.

The motivation here is to have as much of the javascript
as possible into the Qt source tree. This will make
it easier to keep the C++ and js code in sync and
reduce the dependence on a deployment tool/special
javascript on the hosting web page.

In addition add QPepperInstance::postMessage() which
is a thin wrapper over pp::Instance::PostMessage.