Add Qt Quick Playground
The Qt Quick Playground supports live editing QML
in the browser. Several examples are provided [currently
example stubs to be replaced with better content.]

On the C++ side a QML "reloader" class is implemented
that listens on messages from the javascript/web host
with QML source code. A QML root component is crated
on message reception. Parser errors and runtime warnings
are passed back to the web host.

The playground html displays a simple code editor, pre-
populated with QML source code. The source code is sent to
the QML component on changes. Errors and warnings are
presented to the user.

Change-Id: I77f398e26d3b1c26ea312838fe89fe8067d9028f